add getting of refresh token and updating access token
~package.json
- Added library (jwt-decode 3.1.2) that helps decoding JWTs token which are Base64Url encoded.
- Added package 'url-pattern' that helps more easier than regex string matching patterns for urls and other strings. Turn strings into data or data into strings.
***
~server/db/refreshTokens.js
- Added a new request to DB with find unique refresh token by token
***
~server/db/users.js
- Added a new request to DB with find unique user by userId
***
~server/utils/jwt.js
- Added two new functions:
	-- decodeRefreshToken: decoding token
	-- decodeAccessToken:  decoding token
***
~server/api/auth/refresh.get.js
That file is request to DB for updating access token.
The file was checking cookies for contain a refresh token
(Otherwise send error).
Then sends request to DB for checking contain a refresh token in DB
(Otherwise send error).
The token is then decoded and the access token could be refreshed.